Move classes between trello and service to fit new dependency direction
