+++ service/src/main/java/net/kemitix/ugiggle/service/AttachmentConverter.java
@@ -1,5 +1,7 @@
 package net.kemitix.ugiggle.service;
 
+import net.kemitix.ugiggle.trello.Attachment;
+
 import java.util.Optional;
 
 public interface AttachmentConverter {
