+++ service/src/main/java/net/kemitix/ugiggle/service/ConversionService.java
@@ -1,5 +1,7 @@
 package net.kemitix.ugiggle.service;
 
+import net.kemitix.ugiggle.trello.Attachment;
+
 import javax.enterprise.context.ApplicationScoped;
 import javax.enterprise.inject.Instance;
 import javax.inject.Inject;
