+++ service/src/main/java/net/kemitix/ugiggle/service/EmailService.java
@@ -1,5 +1,7 @@
 package net.kemitix.ugiggle.service;
 
+import net.kemitix.ugiggle.trello.Attachment;
+
 import javax.mail.MessagingException;
 import java.io.IOException;
