+++ service/src/main/java/net/kemitix/ugiggle/service/ReadingListService.java
@@ -1,7 +1,9 @@
 package net.kemitix.ugiggle.service;
 
+import net.kemitix.ugiggle.trello.TrelloCard;
+
 import java.util.stream.Stream;
 
 public interface ReadingListService {
-    Stream<? extends Submission> getReadingList();
+    Stream<TrelloCard> getReadingList();
 }
