--- service/src/main/java/net/kemitix/ugiggle/service/Submission.java
@@ -1,9 +1,0 @@
-package net.kemitix.ugiggle.service;
-
-import java.util.stream.Stream;
-
-public interface Submission {
-    Stream<Attachment> findAttachments();
-
-    String getName();
-}
