+++ trello/src/main/java/net/kemitix/ugiggle/trello/Attachment.java
@@ -1,4 +1,4 @@
-package net.kemitix.ugiggle.service;
+package net.kemitix.ugiggle.trello;
 
 import java.io.File;
