+++ trello/src/main/java/net/kemitix/ugiggle/trello/AttachmentDirectoryImpl.java
@@ -1,4 +1,4 @@
-package net.kemitix.ugiggle.service;
+package net.kemitix.ugiggle.trello;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
