+++ trello/src/main/java/net/kemitix/ugiggle/trello/TrelloProducers.java
@@ -4,7 +4,6 @@ import com.julienvey.trello.Trello;
 import com.julienvey.trello.TrelloHttpClient;
 import com.julienvey.trello.impl.TrelloImpl;
 import com.julienvey.trello.impl.http.JDKTrelloHttpClient;
-import net.kemitix.ugiggle.service.UGiggleConfig;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.enterprise.inject.Produces;
